% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator}
\alias{bulkSimulator}
\title{Bulk simulation function}
\usage{
bulkSimulator(
  methods,
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_sample = NA,
  simulated_frac = NULL,
  heter_cell_type = NA,
  ncells_perSample = 500,
  min_chunkSize = 5,
  use_chunk = "all",
  colnames_of_subcluster = NA,
  export_cellUsage = F,
  nbulk = 100,
  n.core = 1,
  ...
)
}
\arguments{
\item{methods}{a character vector indicating which bulk simulation methods to use. Use list_bulkSimulator() to check for available method names.}

\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta. Required for 'semi', 'heter', and 'SCDC' methods.}

\item{simulated_frac}{a matrix with pre-defined fraction of different cell types, with samples in rows and cell_types in columns. This argument if required for 'homo', 'semi', 'heter', 'heter_sampleIDfree', 'immunedeconv' methods}

\item{heter_cell_type}{name of the cell_type to maintain the highest level of heterogeneity. It is recommended to set this parameter to the name of the malignant cell types.
This argument is required for 'semi' and 'heter_sampleIDfree' methods}

\item{ncells_perSample}{number of cells to aggregate for each simulated bulk sample. This is an argument required for 'homo', 'semi', 'immunedeconv' and 'SCDC' methods}

\item{min_chunkSize}{minimum number of cells required to construct a particular cell-type component in the simulated bulk, such as requiring at least 20 cells for B cells, at least 20 cells for T cells, and so forth. This is an argument required for 'semi' and 'heter' methods}

\item{use_chunk}{a character indicating which cells to pool together for the a given cell_type. Default='all' other options include 'random'.
When use_chunk = 'all', use all the cells belonging to the same patient for a given cell type to generate the certain cell type component in the simulated bulk;
when use_chunk = 'random', randomly select 50-100\% of the cells belonging to the same patient for a given cell type. This is an argument required for 'semi' and 'heter' methods}

\item{colnames_of_subcluster}{column name that corresponds to subcluster info in scMeta, where subcluster contains sub-clustering information for each cellType.  This is an argument required for 'heter_sampleIDfree' method only.
Set to NA if subclustering information is not available; the function will then generate subclustering information automatically}

\item{export_cellUsage}{a logical variable determining whether to export cell names used to generate the simulated bulk. Default = F. This is an argument is only applicable to 'homo', 'semi', 'heter' and 'heter_sampleIDfree' methods}

\item{nbulk}{number of simulated bulk samples. This argument is only applicable to 'SCDC' methods, as they will need this argument to generate the fraction matrix. Default = 100}

\item{n.core}{number of cores to use for parallel programming. Default = 1. Parallel programming is applicable to 'homo', 'semi' and 'heter' methods.}

\item{...}{additional arguments to be passed to the following functions: bulkSimulator_heter_sampleIDfree(),and bulkSimulator_SCDC()}
}
\value{
a list of simulated bulk objects
}
\description{
Generate a list of bulk simulation objects using user-selected simulation methods
}
\examples{
\dontrun{
# first create a fraction matrix for the simulated bulk
simulated_frac = fracSimulator_Beta(scMeta, n = 100,
                                    colnames_of_sample = 'sampleID',
                                    colnames_of_cellType = 'cell_type',
                                    fixed_cell_type = 'malignant')

bulkSimulator(methods = c('homo', 'semi','heter'),
              scExpr = scExpr,
              scMeta = scMeta,
              colnames_of_cellType = 'cell_type',colnames_of_sample = 'sampleID',

              # pre-defined simulated_frac will be used for 'homo', 'semi, 'heter' and 'immunedeconv' methods
              simulated_frac = simulated_frac,

              # argument for semi/heter_sampleIDfree bulk simulation method
              heter_cell_type = 'malignant',

              # general simulation parameters
              ncells_perSample = 500,
              min_chunkSize = 20,
              use_chunk = 'random',

              n.core = 16)
}
}
