% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator_favilaco}
\alias{bulkSimulator_favilaco}
\title{Bulk simulation using Generator() function from Favilaco et al.}
\usage{
bulkSimulator_favilaco(
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  nbulk = 100,
  pool.size = 100,
  min.percentage = 1,
  max.percentage = 99,
  seed = 24
)
}
\arguments{
\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{a dataframe that stores annotation info of each cells}

\item{colnames_of_cellType}{column name that corresponds to cellType annotation in scMeta}

\item{nbulk}{number of simulated bulk samples}

\item{pool.size}{number of cells to aggregate for each simulated bulk sample}

\item{min.percentage}{minimum percentage of cellType fraction to generate in fraction simulation. Default = 1}

\item{max.percentage}{maximum percentage of cellType fraction to generate in fraction simulation. Default = 99}

\item{seed}{a single value. Default = 24}
}
\value{
a list containing the simulated bulk expression and its associated simulated fraction matrix
}
\description{
Bulk simulation using Generator() function from Favilaco et al.
}
