% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator_heter}
\alias{bulkSimulator_heter}
\title{Heterogeneous bulk simulation}
\usage{
bulkSimulator_heter(
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_sample = NA,
  simulated_frac = NULL,
  min_chunkSize = 5,
  use_chunk = "all",
<<<<<<< HEAD
  export_cellUsage = FALSE,
=======
  export_cellUsage = F,
>>>>>>> 7ff7b730a10aa03a7aba7cf0128d54a5c25083f8
  n.core = 1
)
}
\arguments{
\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta}

\item{simulated_frac}{a matrix with pre-defined fraction of different cell types, with samples in rows and cell_types in columns}

\item{min_chunkSize}{minimum number of cells required to construct a particular cell-type component in the simulated bulk, such as requiring at least 20 cells for B cells, at least 20 cells for T cells, and so forth.}

\item{use_chunk}{a character indicating which cells to pool together for the heter_cell_type. Default='all' other options include 'random'.
When use_chunk = 'all', use all the cells belonging to the same patient for a given cell type to generate the certain cell type component in the simulated bulk;
when use_chunk = 'random', randomly select 50-100\% of the cells belonging to the same patient for a given cell type}

\item{export_cellUsage}{a logical variable determining whether to export cell names used to generate the simulated bulk. Default = F}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list containing the simulated bulk expression and its associated simulated fraction matrix
}
\description{
Generate bulk samples by aggregating single cells, with each cell-type component is constrained to cells originating from the same patient.
This requires that sampleID is pre-defined in the scRNA dataset.
}
