% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator_heter_sampleIDfree}
\alias{bulkSimulator_heter_sampleIDfree}
\title{sampleID-independent heterogeneous bulk simulation}
\usage{
bulkSimulator_heter_sampleIDfree(
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_cellState = NA,
  simulated_frac = NULL,
  heter_cell_type = NA,
  dirichlet_cs_par = 0.1
)
}
\arguments{
\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_cellState}{column name that corresponds to cellState in scMeta, where cellState contains sub-clustering information for each cellType}

\item{simulated_frac}{a matrix with pre-defined fraction of different cell types, with samples in rows and cell_types in columns}

\item{heter_cell_type}{name of the cell_type to maintain the highest level of heterogeneity. It is recommended to set this parameter to the name of the malignant cell types}

\item{dirichlet_cs_par}{a numeric value determine the dispersion level of the simulated fractions. With lower value indicating higher dispersion level. Default = 0.1}
}
\value{
a list containing the simulated bulk expression and its associated simulated fraction matrix
}
\description{
Generate bulk samples by aggregating cells from sub-clusters of each cellType. The simulated bulk sample comprises a single sub-cluster component
for the heter_cell_type, with heter_cell_type typically recommended to be set as the malignant cell type. For the remaining cell type components, it includes random sub-cluster information
}
