% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator_homo}
\alias{bulkSimulator_homo}
\title{Homogeneous bulk simulation}
\usage{
bulkSimulator_homo(
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  simulated_frac = NULL,
  ncells_perSample = 500,
  n.core = 1
)
}
\arguments{
\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{simulated_frac}{a matrix with pre-defined fraction of different cell types, with samples in rows and cell_types in columns}

\item{ncells_perSample}{number of cells to pool together to generate a simulated bulk sample. Default 500}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list containing the simulated bulk expression and its associated simulated fraction matrix
}
\description{
Generate bulk samples by aggregating single cells with pre-defined fractions
}
