% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator_immunedeconv}
\alias{bulkSimulator_immunedeconv}
\title{Bulk simulation using make_bulk_eset() function from immunedeconv package}
\usage{
bulkSimulator_immunedeconv(
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  simulated_frac = NULL,
  n_cells = 500
)
}
\arguments{
\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{a dataframe that stores annotation info of each cells}

\item{colnames_of_cellType}{column name that corresponds to cellType annotation in scMeta}

\item{simulated_frac}{a matrix with pre-defined fraction of different cell types, with samples in rows and cell_types in columns}

\item{n_cells}{number of cells to aggregate for each simulated bulk sample}
}
\value{
a list containing the simulated bulk expression and its associated simulated fraction matrix
}
\description{
Bulk simulation using make_bulk_eset() function from immunedeconv package
}
