% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator_SCDC}
\alias{bulkSimulator_SCDC}
\title{Bulk simulation using generateBulk_norep() function from SCDC package}
\usage{
bulkSimulator_SCDC(
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_sample = NA,
  nbulk = 10,
  disease = NULL,
  ct.sub = NULL,
  prop_mat = NULL,
  samplewithRep = T
)
}
\arguments{
\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{a dataframe that stores annotation info of each cells}

\item{colnames_of_cellType}{column name that corresponds to cellType annotation in scMeta}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta}

\item{nbulk}{number of simulated bulk samples}

\item{disease}{indicate the health condition of subjects}

\item{ct.sub}{a subset of cell types that are selected to construct pseudo bulk samples. If NULL, then all cell types are used.}

\item{prop_mat}{manually input the cell-type proportion for pseudo bulk samples; with cellTypes in columns and samples in rows}

\item{samplewithRep}{logical, randomly sample single cells with replacement. Default is T}
}
\value{
a list containing the simulated bulk expression and its associated simulated fraction matrix
}
\description{
Bulk simulation using generateBulk_norep() function from SCDC package
}
