% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator_semi}
\alias{bulkSimulator_semi}
\title{Semi-heterogeneous bulk simulation}
\usage{
bulkSimulator_semi(
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_sample = NA,
  heter_cell_type = NA,
  simulated_frac = NULL,
  ncells_perSample = 500,
  min_chunkSize = 10,
  use_chunk = "all",
  export_cellUsage = F,
  n.core = 1
)
}
\arguments{
\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta}

\item{heter_cell_type}{cell_type to maintain heterogeneity across samples}

\item{simulated_frac}{a matrix with pre-defined fraction of different cell types, with samples in rows and cell_types in columns}

\item{ncells_perSample}{number of cells to pool together to generate a simulated bulk sample. Default 500}

\item{min_chunkSize}{minimum number of cells required to construct the heter_cell_type}

\item{use_chunk}{a character indicating which cells to pool together for the heter_cell_type. Default='all' other options include 'random'
When use_chunk = 'all', use all the cells belonging to the same patient for a given cell type to generate the certain cell type component in the simulated bulk;
when use_chunk = 'random', randomly select 50-100\% of the cells belonging to the same patient for a given cell type}

\item{export_cellUsage}{a logical variable determining whether to export cell names used to generate the simulated bulk. Default = F}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list containing the simulated bulk expression and its associated simulated fraction matrix
}
\description{
Generate bulk samples by aggregating single cells, while maintaining heterogeneity for a certain cell type.
Specifically, heter_cell_type is constrained to cells originating from a single patient. This requires that sampleID is pre-defined
in the scRNA dataset
}
