% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{create_train_test_splitting}
\alias{create_train_test_splitting}
\title{Training/testing cells splitting}
\usage{
create_train_test_splitting(
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_sample = NA,
  training_ratio = 0.5,
  split_by = "sample"
)
}
\arguments{
\item{scMeta}{single cell annotation file with cells as rownames}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta}

\item{training_ratio}{a numeric value indicating ratio of training cells. Default = 0.5}

\item{split_by}{a string vector indicating how to split training and testing cells. Available options includes 'cell' and 'sample'.
When split_by = 'cell', for each cell type, a pre-defined ratio of cells will be sampled as training cells. When split = 'sample',
a pre-defined ratio of samples will be sampled as training samples, and all the cells associated will become training cells.}
}
\value{
a list of training cells and testing cells
}
\description{
Training/testing cells splitting
}
\examples{
\dontrun{
# when splitting by cells
create_train_test_splitting(scMeta = scMeta, colnames_of_cellType = 'cell_type',
                           training_ratio = 0.5, split_by = 'cell')

# when splitting by samples
create_train_test_splitting(scMeta = scMeta, colnames_of_sample = 'sampleID',
                           training_ratio = 0.5, split_by = 'sample')
}
}
