% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_Bayesian.R
\name{deconv_Bayesian}
\alias{deconv_Bayesian}
\title{Bayesian-based deconvolution}
\usage{
deconv_Bayesian(
  methods,
  bulk_expr,
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_cellState = NA,
  colnames_of_sample = NA,
  key = NA,
  n.iter = 100,
  n.core = 1
)
}
\arguments{
\item{methods}{a character vector indicating which methods to use. Use list_deconv_Bayesian() to check for available method names.}

\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{scExpr}{scRNA expression to build the reference, with genes in rows and samples in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_cellState}{column name that corresponds to cellState in scMeta. Set to NA if not available}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta}

\item{key}{InstaPrism argument: name of the malignant cell type. Upon setting the key parameter, the updated malignant reference
will be unique for each individual. Set to NA if there is no malignant cells in the problem, and the updated reference will be the same for all the individuals}

\item{n.iter}{InstaPrism argument: number of iterations. Default = 100}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of Bayesian-based deconvolution results
}
\description{
Bayesian-based deconvolution
}
\examples{
\dontrun{
deconv_Bayesian(methods = c('InstaPrism'),
                bulk_expr = bulk_expr,
                scExpr = scExpr,
                scMeta = scMeta,
                colnames_of_cellType = 'cell_type',
                colnames_of_sample = 'sampleID',
                key = 'malignant',
                n.core = 4)

}
}
