% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_Bayesian.R
\name{deconv_Bayesian_BayesPrism}
\alias{deconv_Bayesian_BayesPrism}
\title{Bayesian-based deconvolution with BayesPrism}
\usage{
deconv_Bayesian_BayesPrism(
  bulk_expr,
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_cellState = NA,
  colnames_of_sample = NA,
  outlier.cut = 0.01,
  outlier.fraction = 0.1,
  key = NA,
  n.core = 1
)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{scExpr}{scRNA expression to build the reference, with genes in rows and samples in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_cellState}{column name that corresponds to cellState in scMeta. Set to NA if this information is not available.
In this case, the function will utilize cross-sample heterogeneity and assign cell states according to their original sampleID
for the key cell type. For the remaining cell types, the cell state will be the same as the cell type.}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta. This argument will be called when colnames_of_cellState is not available}

\item{outlier.cut}{Filter genes in bulk mixture whose expression fraction is greater than outlier.cut (Default=0.01)
in more than outlier.fraction (Default=0.01) of bulk data. Removal of outlier genes will ensure that the inference will not be dominated by outliers.}

\item{outlier.fraction}{Filter genes in bulk mixture whose expression fraction is greater than outlier.cut (Default=0.01)
in more than outlier.fraction (Default=0.1) of bulk data. Removal of outlier genes will ensure that the inference will not be dominated by outliers.}

\item{key}{BayesPrism argument: name of the malignant cell type. Upon setting the key parameter, the updated malignant reference
will be unique for each individual. Set to NA if there is no malignant cells in the problem, and the updated reference will be the same for all the individuals}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of BayesPrism deconvolution results using both initial reference and updated reference.
}
\description{
Bayesian-based deconvolution with BayesPrism
}
