% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_Bayesian.R
\name{deconv_Bayesian_InstaPrism}
\alias{deconv_Bayesian_InstaPrism}
\title{Bayesian-based deconvolution with InstaPrism}
\usage{
deconv_Bayesian_InstaPrism(
  bulk_expr,
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_cellState = NA,
  colnames_of_sample = NA,
  key = NA,
  n.iter = 100,
  n.core = 1
)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{scExpr}{scRNA expression to build the reference, with genes in rows and samples in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_cellState}{column name that corresponds to cellState in scMeta. Set to NA if not available}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta}

\item{key}{InstaPrism argument: name of the malignant cell type. Upon setting the key parameter, the updated malignant reference
will be unique for each individual. Set to NA if there is no malignant cells in the problem, and the updated reference will be the same for all the individuals}

\item{n.iter}{InstaPrism argument: number of iterations. Default = 100}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results using both initial reference and updated reference.  If the 'key' argument is valid,
it also includes the results using the malignant_updated reference (with malignant updated only)
}
\description{
Bayesian-based deconvolution with InstaPrism
}
