% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_marker.R
\name{deconv_marker}
\alias{deconv_marker}
\title{Marker-based deconvolution}
\usage{
deconv_marker(
  methods,
  bulk_expr,
  marker_list,
  alpha = NULL,
  sigma = NULL,
  epsilon = 0.001,
  maxIter = 1000,
  n.core = 1
)
}
\arguments{
\item{methods}{a character vector indicating which methods to use. Use list_deconv_marker() to check for available method names.}

\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{marker_list}{a list of cell-type specific markers. Each list element is named by the corresponding method used to generate the markers.}

\item{alpha}{TOAST::MDeconv() argument. A vector including the prior mean for all cell types. Set to NULL to for TOAST/-P (partial reference-free deconvolution without prior)}

\item{sigma}{TOAST::MDeconv() argument. A vector including the prior standard deviation for all cell types.  Set to NULL to for TOAST/-P (partial reference-free deconvolution without prior)}

\item{epsilon}{TOAST::MDeconv() argument. A numeric variable to control the level of convergence. With a large epsilon, it converges quickly and the algorithm may not converge well.
With a small epsilon, it converges slower and the algorithm may converge "too much". The default value is 1e-3, which we find is a reasonable threshold.}

\item{maxIter}{TOAST::MDeconv() argument. Number of maximum iterations.}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different input markers and different marker-based methods. The results are named using the format
'MarkerBased_' + marker-based method, followed by the name of the respective input markers
}
\description{
Generate a list of marker-based deconvolution results with user-selected methods
}
\examples{
\dontrun{
# create marker list with refMarkers() function
marker_list = refMarkers(methods = c('limma','scran'),
              scExpr = scExpr,
              cell_type_labels = scMeta$cell_type)

deconv_marker(methods = c('firstPC','gsva','debCAM','TOAST'),
              bulk_expr = bulk_expr,
              marker_list = marker_list)
}
}
