% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_marker.R
\name{deconv_markers_gsva}
\alias{deconv_markers_gsva}
\title{Marker-based deconvolution with GSVA}
\usage{
deconv_markers_gsva(bulk_expr, marker_list, n.core = 1)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{marker_list}{a list of cell-type specific markers. Each list element is named by the corresponding method used to generate the markers.}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different input markers. The results are named using the format
'MarkerBased_gsva_' followed by the name of the method used to generate the markers, such as 'limma'
}
\description{
Calculate relative enrichment of cell-type specific markers using ssGSVA. Please note that the resulting score
represents an abundance estimate and should not be interpreted as fractions.
}
