% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_marker.R
\name{deconv_markers_TOAST}
\alias{deconv_markers_TOAST}
\title{Marker-based deconvolution with TOAST::MDeconv()}
\usage{
deconv_markers_TOAST(
  bulk_expr,
  marker_list,
  alpha = NULL,
  sigma = NULL,
  epsilon = 0.001,
  maxIter = 1000,
  n.core = 1
)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{marker_list}{a list of cell-type specific markers. Each list element is named by the corresponding method used to generate the markers.}

\item{alpha}{TOAST::MDeconv() argument. A vector including the prior mean for all cell types. Set to NULL to for TOAST/-P (partial reference-free deconvolution without prior)}

\item{sigma}{TOAST::MDeconv() argument. A vector including the prior standard deviation for all cell types.  Set to NULL to for TOAST/-P (partial reference-free deconvolution without prior)}

\item{epsilon}{TOAST::MDeconv() argument. A numeric variable to control the level of convergence. With a large epsilon, it converges quickly and the algorithm may not converge well.
With a small epsilon, it converges slower and the algorithm may converge "too much". The default value is 1e-3, which we find is a reasonable threshold.}

\item{maxIter}{TOAST::MDeconv() argument. Number of maximum iterations.}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different input markers. The results are named using the format
'MarkerBased_TOAST_' followed by the name of the method used to generate the markers, such as 'limma'
}
\description{
Perform bulk deconvolution using MDeconv() function from TOAST package. The resulting scores have sum-to-one constraint and
can be interpreted as fractions.
}
