% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_refFree.R
\name{deconv_refFree}
\alias{deconv_refFree}
\title{Reference free deconvolution}
\usage{
deconv_refFree(
  methods,
  bulk_expr,
  k,
  corner.strategy = 2,
  dim.rdc = 10,
  thres.low = 0.05,
  thres.high = 0.95,
  cluster.method = "K-Means",
  cluster.num = 50,
  MG.num.thres = 20,
  lof.thres = 0.02,
  quickhull = TRUE,
  quick.select = NULL,
  sample.weight = NULL,
  appro3 = TRUE,
  generalNMF = FALSE,
  cores = NULL
)
}
\arguments{
\item{methods}{a character vector indicating which methods to use. Use list_deconv_refFree() to check for available method names.}

\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{k}{number of cell types in bulk expression}

\item{corner.strategy}{The method to find corners of convex hull. 1: minimum sum of margin-of-errors; 2: minimum sum of reconstruction errors. The default is 2.}

\item{dim.rdc}{Reduced data dimension; should be not less than maximum candidate K.}

\item{thres.low}{The lower bound of percentage of genes to keep for CAM with ranked norm. The value should be between 0 and 1. The default is 0.05.}

\item{thres.high}{The higher bound of percentage of genes to keep for CAM with ranked norm. The value should be between 0 and 1. The default is 0.95.}

\item{cluster.method}{The method to do clustering. The default "K-Means" will use kmeans. The alternative "apcluster" will use apclusterK-methods.}

\item{cluster.num}{The number of clusters; should be much larger than K. The default is 50.}

\item{MG.num.thres}{The clusters with the gene number smaller than MG.num.thres will be treated as outliers. The default is 20.}

\item{lof.thres}{Remove local outlier using lofactor. MG.num.thres is used as the number of neighbors in the calculation of the local outlier factors. The default value 0.02 will remove top 2\% local outliers. Zero value will disable lof.}

\item{quickhull}{Perform quickhull to select clusters or not. The default is True.}

\item{quick.select}{The number of candidate corners kept after quickhull and SFFS greedy search. If Null, only quickhull is applied. The default is 20. If this value is larger than the number of candidate corners after quickhull, greedy search will also not be applied.}

\item{sample.weight}{Vector of sample weights. If NULL, all samples have the same weights. The length should be the same as sample numbers. All values should be positive.}

\item{appro3}{Estimate A and S matrix by approach 3 or not. Please see CAMASest for further information. The default is TRUE.}

\item{generalNMF}{If TRUE, the decomposed proportion matrix has no sum-to-one constraint for each row. The default is FALSE. TRUE value brings two changes: (1) Without assuming samples are normalized,
the first principal component will not forced to be along c(1,1,..,1) but a standard PCA will be applied during preprocessing. (2) Without sum-to-one constraint for each row, the scale ambiguity of each column vector in proportion matrix will not be removed.}

\item{cores}{The number of system cores for parallel computing. If not provided, one core for each element in K will be invoked. Zero value will disable parallel computing.}
}
\value{
a list of deconvolution results obtained using different reference free methods. The results are named using the format 'refFree_' + method
}
\description{
Reference free deconvolution
}
\examples{
\dontrun{
deconv_refFree(methods = c('linseed','debCAM'),
               bulk_expr = bulk_expr, k = 8)
}
}
