% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_refFree.R
\name{deconv_refFree_linseed}
\alias{deconv_refFree_linseed}
\title{Reference free deconvolution with linseed}
\usage{
deconv_refFree_linseed(bulk_expr, k)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{k}{number of cell types in bulk expression}
}
\value{
a matrix of deconvolution results with cell types in rows and samples in columns
}
\description{
Reference free deconvolution with linseed
}
