% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_regression.R
\name{deconv_regression_cibersort}
\alias{deconv_regression_cibersort}
\title{Regression-based deconvolution with CIBERSORT}
\usage{
deconv_regression_cibersort(
  bulk_expr,
  sigMatrix_list,
  cibersort_path = NULL,
  QN = F,
  skip_raw = T,
  n.core = 1
)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{sigMatrix_list}{a list of signature matrix}

\item{cibersort_path}{path to CIBERSORT.R}

\item{QN}{a logical variable determining whether to quantile normalize the input matrix. Default = F}

\item{skip_raw}{a logical variable indicating whether to skip the signature matrix named 'raw', which corresponds to the raw reference matrix without gene filtering.
Setting this argument to TRUE is recommended to save computation time and resources.}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different input signature matrices. The results are named using the format
'RefBased_cibersort_' followed by the name of the respective input signature matrix.
}
\description{
Perform bulk deconvolution using the cibersort algorithm with input signature matrices
}
