% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_regression.R
\name{deconv_regression_MuSiC}
\alias{deconv_regression_MuSiC}
\title{Regression-based deconvolution with MuSiC}
\usage{
deconv_regression_MuSiC(
  bulk_expr,
  sigMatrix_list,
  scExpr = NULL,
  scMeta = NULL,
  colnames_of_cellType = NA,
  colnames_of_sample = NA,
  normalize = F,
  n.core = 1
)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{sigMatrix_list}{a list of signature matrix}

\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta}

\item{normalize}{MuSiC::music_prop() normalize argument. Default = F}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different marker genes from input signature matrices. The results are named using the format
'RefBased_MuSiC_' followed by the name of the respective input signature matrix.
}
\description{
Perform bulk deconvolution using MuSiC with genes selected from the input signature matrices as MuSiC input markers
}
