% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_regression.R
\name{deconv_regression_nnls}
\alias{deconv_regression_nnls}
\title{Regression-based deconvolution with nnls}
\usage{
deconv_regression_nnls(bulk_expr, sigMatrix_list, n.core = 1)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{sigMatrix_list}{a list of signature matrix}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different input signature matrices. The results are named using the format
'RefBased_nnls_' followed by the name of the respective input signature matrix.
}
\description{
Perform bulk deconvolution using the Non-Negative Least Squares (NNLS) algorithm with input signature matrices
}
