% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_regression.R
\name{deconv_regression_RPC}
\alias{deconv_regression_RPC}
\title{Regression-based deconvolution with robust partial correlations}
\usage{
deconv_regression_RPC(
  bulk_expr,
  sigMatrix_list,
  skip_raw = T,
  maxit = 100,
  n.core = 1
)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{sigMatrix_list}{a list of signature matrix}

\item{skip_raw}{a logical variable indicating whether to skip the signature matrix named 'raw', which corresponds to the raw reference matrix without gene filtering.
Setting this argument to TRUE is recommended to save computation time and resources.}

\item{maxit}{EpiDISH::epidish() argument, an integer indicating the limit of the number of IWLS iterations. Default = 100}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different input signature matrices. The results are named using the format
'RefBased_RPC_' followed by the name of the respective input signature matrix.
}
\description{
Perform bulk deconvolution using robust partial correlations (RPC) algorithm from EpiDISH package, with the input signature matrices
}
