% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_regression.R
\name{deconv_regression_wRLM}
\alias{deconv_regression_wRLM}
\title{Regression-based deconvolution with weighted robust linear regression}
\usage{
deconv_regression_wRLM(
  bulk_expr,
  sigMatrix_list,
  skip_raw = TRUE,
  weight = TRUE,
  intercept = TRUE,
  scale = FALSE,
  QN = FALSE,
  verbose = FALSE,
  n.core = 1
)
}
\arguments{
\item{bulk_expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{sigMatrix_list}{a list of signature matrix}

\item{skip_raw}{a logical variable indicating whether to skip the signature matrix named 'raw', which corresponds to the raw reference matrix without gene filtering.
Setting this argument to TRUE is recommended to save computation time and resources.}

\item{weight}{LinDeconSeq::deconSeq() argument, a logical variable determining whether to weight gene or not. Default = T}

\item{intercept}{LinDeconSeq::deconSeq() argument, a logical variable determining whether to add intercept when using robust linear regression. Default = T}

\item{scale}{LinDeconSeq::deconSeq() argument, a logical variable determining whether to scale bulk gene expression or not. Default = T}

\item{QN}{LinDeconSeq::deconSeq() argument, a logical variable determining whether to quantile normalize bulk expression profile. Default = T}

\item{verbose}{LinDeconSeq::deconSeq() argument, a logical variable determining whether to print detailed information. Default = F}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list of deconvolution results obtained using different input signature matrices. The results are named using the format
'RefBased_wRLM_' followed by the name of the respective input signature matrix.
}
\description{
Perform bulk deconvolution using weighted robust linear regression (w-RLM) algorithm from LinDeconSeq package, with the input signature matrices
}
