% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_simulation.R
\name{fracSimulator_Beta}
\alias{fracSimulator_Beta}
\title{Simulate realistic cell-type fractions from beta distribution}
\usage{
fracSimulator_Beta(
  scMeta,
  n,
  colnames_of_sample = NA,
  colnames_of_cellType = NA,
  fixed_cell_type = NA,
  min.frac = 0,
  showFractionPlot = T
)
}
\arguments{
\item{scMeta}{a dataframe that stores annotation info of each cells}

\item{n}{number of simulated bulk samples}

\item{colnames_of_sample}{column name that corresponds to cellType in scMeta}

\item{colnames_of_cellType}{column name that corresponds to sampleID in scMeta}

\item{fixed_cell_type}{a character denotes the target cell type for which we strive to faithfully preserve its distribution.
It is recommended to set this parameter to the name of the malignant cell types. If left undefined, the function will automatically
select the most abundant cell type as 'fixed_cell_type'.}

\item{min.frac}{minimum fraction in the simulated fraction, values below this threshold will be set to zero}

\item{showFractionPlot}{a logical variable determining whether to display simulated fraction distribution for the fixed_cell_type}
}
\value{
a matrix of simulated fraction, with samples in rows and cell_types in columns
}
\description{
Simulate cell-type fractions by utilizing the Cell-Type Proportion Distribution derived from scRNA.
This approach is recommended when the scMeta contains a sufficient number of distinct samples to accurately capture the distribution
and generate realistic proportions (n >= 10 is recommended).
}
\examples{
\dontrun{
fracSimulator_Beta(scMeta, n = 100, colnames_of_sample = 'sampleID',
                   colnames_of_cellType = 'cell_type', fixed_cell_type = 'malignant')
}
}
