% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_simulation.R
\name{fracSimulator_Dirichlet}
\alias{fracSimulator_Dirichlet}
\title{Simulate cell-type fractions from dirichlet distribution}
\usage{
fracSimulator_Dirichlet(
  frac_prior,
  n,
  dispersion_par = 0.1,
  min.frac = 0,
  showFractionPlot = T
)
}
\arguments{
\item{frac_prior}{A named numeric vector representing the fraction prior. The simulated fraction will fluctuate around the values specified in this vector.}

\item{n}{number of random fractions to generate}

\item{dispersion_par}{a numeric value determine the dispersion level of the simulated fractions. With lower value indicating higher dispersion level.}

\item{min.frac}{minimum fraction in the simulated fraction, values below this threshold will be set to zero}

\item{showFractionPlot}{a logical variable determining whether to display the boxplot of simulated fractions. It is recommended to set this parameter to TRUE to
visualize the effect dispersion_par and choose the optimal value of it}
}
\value{
a matrix of simulated fraction, with samples in rows and cell_types in columns
}
\description{
Simulate cell-type fractions from dirichlet distribution
}
\examples{
frac_prior = c(100,200,300,400)
names(frac_prior) = paste0('cellType',seq(1,4))
fracSimulator_Dirichlet(frac_prior, n = 10, dispersion_par = 0.01)
}
