% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{gather_performance}
\alias{gather_performance}
\title{Summarize deconvolution performance in a long table}
\usage{
gather_performance(deconvPerformance)
}
\arguments{
\item{deconvPerformance}{a deconvPerformance object returned from benchmarking_evalu() function}
}
\value{
a dataframe summarizing the deconvolution performance per deconvolution method and simulation method
}
\description{
Summarize deconvolution performance in a long table
}
