% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExplorer.R
\name{list_bulkSimulator}
\alias{list_bulkSimulator}
\title{List available bulk simulation strategies}
\usage{
list_bulkSimulator(show_description = F)
}
\arguments{
\item{show_description}{a logical variable to determine whether to show full description of each method}
}
\value{
a dataframe of available bulk simulation strategies along with the corresponding function names to execute these methods.
}
\description{
List available bulk simulation strategies
}
\examples{
list_bulkSimulator()
}
