% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExplorer.R
\name{list_deconv_Bayesian}
\alias{list_deconv_Bayesian}
\title{List available deconvolution methods using Bayesian framework}
\usage{
list_deconv_Bayesian(show_description = F)
}
\arguments{
\item{show_description}{a logical variable to determine whether to show full description of each method}
}
\value{
a dataframe of available Bayesian-based deconvolution methods along with the corresponding function names to execute these methods.
}
\description{
List available deconvolution methods using Bayesian framework
}
\examples{
list_deconv_Bayesian()
}
