% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExplorer.R
\name{list_deconv_marker}
\alias{list_deconv_marker}
\title{List available methods to perform marker-based deconvolution}
\usage{
list_deconv_marker(show_description = F)
}
\arguments{
\item{show_description}{a logical variable to determine whether to show full description of each method}
}
\value{
a dataframe of available marker-based deconvolution methods along with the corresponding function names to execute these methods.
}
\description{
List available methods to perform marker-based deconvolution
}
\examples{
list_deconv_marker()
}
