% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExplorer.R
\name{list_fracSimulator}
\alias{list_fracSimulator}
\title{List available fraction simulation strategies}
\usage{
list_fracSimulator(show_description = F)
}
\arguments{
\item{show_description}{a logical variable to determine whether to show full description of each method}
}
\value{
a dataframe of available fraction simulation strategies along with the corresponding function names to execute these methods.
}
\description{
List available fraction simulation strategies
}
\examples{
list_fracSimulator()
}
