% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExplorer.R
\name{list_refMarix}
\alias{list_refMarix}
\title{List available methods to construct signature matrix from scRNA reference}
\usage{
list_refMarix(show_description = F)
}
\arguments{
\item{show_description}{a logical variable to determine whether to show full description of each method}
}
\value{
a dataframe of available methods to construct signature matrix and the corresponding function names to execute these methods
}
\description{
List available methods to construct signature matrix from scRNA reference
}
\examples{
list_refMarix()
}
