% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodExplorer.R
\name{list_refMarkers}
\alias{list_refMarkers}
\title{List available fraction to get cell-type markers from scRNA reference}
\usage{
list_refMarkers(show_description = F)
}
\arguments{
\item{show_description}{a logical variable to determine whether to show full description of each method}
}
\value{
a dataframe of available methods to get markers and the corresponding function names to execute these methods
}
\description{
List available fraction to get cell-type markers from scRNA reference
}
\examples{
list_refMarkers()
}
