% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{plot_MeanVariance}
\alias{plot_MeanVariance}
\title{Mean variance plot}
\usage{
plot_MeanVariance(
  simulator_Statistics_res,
  genes_to_evalu = NULL,
  desired_order = NULL,
  ylim = c(0, 20),
  ybreaks = seq(0, 20, 5),
  ...
)
}
\arguments{
\item{simulator_Statistics_res}{a list of summary statistics returned from simulator_Statistics() function}

\item{genes_to_evalu}{genes to visualize. If set to NULL, will compare all the genes in simulator_Statistics_res}

\item{desired_order}{a character vector indicating desired order to organize simulated bulk in the figure. Default = NULL, in which case figures will be ordered alphabetically.}

\item{ylim}{scale_y_continuous() "limits" parameter. Default = c(0,20)}

\item{ybreaks}{scale_y_continuous() "breaks" parameter. Default = c(0,20)}

\item{...}{additional arguments to pass to theme()}
}
\value{
a panel of mean-variance plot
}
\description{
Mean variance plot
}
