% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{plot_PairwiseSampleCorrelation}
\alias{plot_PairwiseSampleCorrelation}
\title{Plot sample wise correlations in simulated bulk and real bulk}
\usage{
plot_PairwiseSampleCorrelation(
  simulator_Statistics_res,
  desired_order = c("baseline", "homo", "semi", "heter", "heter_sampleIDfree",
    "favilaco", "immunedeconv", "SCDC"),
  ...
)
}
\arguments{
\item{simulator_Statistics_res}{a list of summary statistics returned from simulator_Statistics() function}

\item{desired_order}{a character vector indicating desired order to organize simulated bulk in the figure.
The default order is specified as c('baseline','homo','semi','heter','heter_sampleIDfree','favilaco','immunedeconv','SCDC')}

\item{...}{additional arguments to pass to geom_violin()}
}
\value{
a panel of violin plot showing the distribution of sample-wise correlations
}
\description{
Plot sample wise correlations in simulated bulk and real bulk
}
