% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{plot_variance_comparison}
\alias{plot_variance_comparison}
\title{Comparison of variance: simulated bulk vs real bulk}
\usage{
plot_variance_comparison(
  simulator_Statistics_res,
  genes_to_evalu = NULL,
  variance_type = "cv",
  nrow_panel = 1,
  axis_limit = 5,
  desired_order = c("homo", "semi", "heter", "heter_sampleIDfree", "favilaco",
    "immunedeconv", "SCDC"),
  ...
)
}
\arguments{
\item{simulator_Statistics_res}{a list of summary statistics returned from simulator_Statistics() function}

\item{genes_to_evalu}{genes to compare. If set to NULL, will compare all the genes in simulator_Statistics_res}

\item{variance_type}{a string indicating type of variance to compare. Available options including "cv" and "MAD". Deafult = "cv"}

\item{nrow_panel}{nrow of the ggplot panels}

\item{axis_limit}{axis_limit in the variance comparison plot}

\item{desired_order}{a character vector indicating desired order to organize simulated bulk in the figure.
The default order is specified as c('homo','semi','heter','heter_sampleIDfree','favilaco','immunedeconv','SCDC')}

\item{...}{additional arguments to pass to ggpointdensity()}
}
\value{
a panel of ggpointdensity plots comparing variance in each simulated bulk with real bulk expression
}
\description{
Comparison of variance: simulated bulk vs real bulk
}
