% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMatrix.R
\name{pre_refMatrix_autogeneS}
\alias{pre_refMatrix_autogeneS}
\title{Prepare input for autogeneS}
\usage{
pre_refMatrix_autogeneS(
  scExpr,
  cell_type_labels,
  hv_genes = NULL,
  max.spec_cutoff_for_autogeneS = 0.5,
  autogeneS_input_file_name = NULL,
  display_autogeneS_command = T,
  ngen = 5000,
  seed = 0,
  nfeatures = 400,
  mode = "fixed"
)
}
\arguments{
\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}

\item{hv_genes}{a character vector containing the names of high-variable genes. scExpr will be pre-filtered based on the provided hv_genes to reduce computation time during the differential expression (DE) analysis.
If set to NULL, the function will automatically select genes with specificity score passing 'max.spec' threshold as hv_genes}

\item{max.spec_cutoff_for_autogeneS}{specificity score threshold to select for hv_genes. Default = 0.5}

\item{autogeneS_input_file_name}{desired file name to save the processed file}

\item{display_autogeneS_command}{a logical variable indicating whether to display the command lines to run autogeneS. By pasting the generated code into the command line, an external Python file will be executed,
which will return the generated signature matrices in the 'autogeneS_output' folder.}

\item{ngen}{a numeric variable indicating number of generations used in autogeneS}

\item{seed}{autogeneS seed argument. Default = 0}

\item{nfeatures}{autogeneS nfeatures argument. Default = 400}

\item{mode}{autogeneS mode argument. Default = 'fixed'}
}
\value{
save the filtered signature matrix in the 'autogeneS_input' folder for further signature matrix construction with autogeneS
}
\description{
This function calculates the mean expression across cell types to form a signature matrix. It then filters the signature matrix by selecting highly variable genes
to ensure computational efficiency; finally, it saves the resulting signature matrix in the 'autogeneS_input' folder for further signature matrix construction with autogeneS.
}
