% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMatrix.R
\name{pre_refMatrix_cibersortx}
\alias{pre_refMatrix_cibersortx}
\title{Prepare input for cibersortx server}
\usage{
pre_refMatrix_cibersortx(
  scExpr,
  cell_type_labels,
  downsample = F,
  downsample_ratio = 0.2,
  scMeta = NULL,
  colnames_of_sample = NA,
  colnames_of_cellType = NA,
  cibersortx_input_file_name = NULL
)
}
\arguments{
\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}

\item{downsample}{a logical variable indicating whether to downsample the processed file of not. This argument is useful when the processed file exceeds the storage limit for the cibersortx server}

\item{downsample_ratio}{a numeric value indicating downsampling ratio. This argument determines the fraction of the original size that will be retained in the downsampling process}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta. Required for 'semi', 'heter', and 'SCDC' methods.}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{cibersortx_input_file_name}{desired file name to save the processed file}
}
\value{
save the processed file in 'cibersortx_input' folder
}
\description{
This function pre-processes the single-cell RNA profile and saves the processed file in the 'cibersortx_input' folder. The saved file is in the correct format
to be used as input for constructing a signature matrix with the cibersortx online portal
}
