% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMarkers.R
\name{refMarkers}
\alias{refMarkers}
\title{Obtain cell-type specific markers from scRNA reference}
\usage{
refMarkers(
  methods,
  scExpr = NULL,
  cell_type_labels = NULL,
  cell_state_labels = NULL,
  hv_genes = NULL,
  log2FC = 2,
  log2FC_flexible = 1,
  minimum_n = 15,
  maximum_n = 50,
  max.spec_cutoff_for_DE = 0.3,
  sigMatrixList = NULL
)
}
\arguments{
\item{methods}{a character vector indicating which methods to use. Use list_refMarkers() to check for available method names.}

\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}

\item{cell_state_labels}{a character vector indicating cell-states of each cells in scExpr. Pairwise DE analysis is initially conducted between cell states and then merges for cell types.
If set to NULL, it will default to the same values as the cell_type_labels. Required for 'scran' method}

\item{hv_genes}{a character vector containing the names of high-variable genes. scExpr will be pre-filtered based on the provided hv_genes to reduce computation time during the differential expression (DE) analysis.
If set to NULL, the function will automatically select genes with specificity score passing 'max.spec_cutoff_for_DE' threshold as hv_genes}

\item{log2FC}{log fold change threshold to select marker genes. Marker genes will be limited to a maximum of 'maximum_n' genes among those that pass the 'log2FC' threshold.}

\item{log2FC_flexible}{a flexible log fold change threshold to select marker genes. If there are fewer than 'minimum_n' genes that pass the 'log2FC_flexible' threshold,
all the genes that pass the threshold will be considered as marker genes.}

\item{minimum_n}{minimum number of marker genes for a cell-type}

\item{maximum_n}{maximum number of marker genes of a cell-type}

\item{max.spec_cutoff_for_DE}{specificity score threshold to select for hv_genes. Default = 0.3}

\item{sigMatrixList}{a list of signature matices to derive markers from. This argument is required for 'sigMatrixList' method}
}
\value{
a list of cell-type specific markers generated from use-defined methods
}
\description{
Generate a list of markers from scRNA reference with user-selected methods
}
\examples{
\dontrun{
# Obtain cell-type specific markers with DE analysis
refMarkers(methods = c('limma','scran'), scExpr = scExpr, cell_type_labels = scMeta$cell_type)

# Obtain cell-type specific markers from a list of signature matrices
# In real practice, these signature matrices can be generated using various methods outside of the provided DE-based methods, such as from the cibersortx server or using autogeneS.
# For more details on how to prepare the scRNA data for cibersortx and autogeneS, use the following command:
help("pre_refMatrix_autogeneS")
help("pre_refMatrix_cibersortx")
refMarkers(method = 'sigMatrixList', sigMatrixList = sigMatrixList, maximum_n = 50)
}
}
