% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMarkers.R
\name{refMarkers_scran}
\alias{refMarkers_scran}
\title{Obtain cell-type specific markers with scran DE analysis}
\usage{
refMarkers_scran(
  scExpr,
  cell_type_labels,
  cell_state_labels = NULL,
  hv_genes = NULL,
  log2FC = 2,
  minimum_n = 15,
  maximum_n = 50,
  order_by = "pval",
  max.spec_cutoff_for_DE = 0.3
)
}
\arguments{
\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}

\item{cell_state_labels}{a character vector indicating cell-states of each cells in scExpr. Pairwise DE analysis is initially conducted between cell states and then merges for cell types.
If set to NULL, it will default to the same values as the cell_type_labels.}

\item{hv_genes}{a character vector containing the names of high-variable genes. scExpr will be pre-filtered based on the provided hv_genes to reduce computation time during the differential expression (DE) analysis.
If set to NULL, the function will automatically select genes with specificity score passing 'max.spec_cutoff_for_DE' threshold as hv_genes}

\item{log2FC}{log fold change threshold to select marker genes. Marker genes will be limited to a maximum of 'maximum_n' genes among those that pass the 'log2FC' threshold.}

\item{minimum_n}{minimum number of marker genes for a cell-type. If a cell type has fewer than 'minimum_n' genes passing the log2FC threshold, it will be excluded from the marker list}

\item{maximum_n}{maximum number of marker genes of a cell-type}

\item{order_by}{a character specifying the criteria for ordering DE genes. Available options include 'pval' and 'min.lfc'. When more than 'maximum_n' genes pass the logFC threshold,
markers will be selected from the top 'maximum_n' genes ordered by this criterion. Default = 'pval'.}

\item{max.spec_cutoff_for_DE}{specificity score threshold to select for hv_genes. Default = 0.3}
}
\value{
a list of cell-type marker genes
}
\description{
Obtain cell-type specific markers by by conducting differential expression (DE) analysis using the scran package, which utilizes
pairwise comparison between each cell-types and filters for marker genes by selecting DE genes that meet a specified logFC threshold
}
