% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMarkers.R
\name{refMarkers_Seurat}
\alias{refMarkers_Seurat}
\title{Obtain cell-type specific markers using Seurat::FindAllMarkers()}
\usage{
refMarkers_Seurat(
  scExpr,
  cell_type_labels,
  seurat_marker_method = "wilcox",
  log2FC = 2,
  minimum_n = 15,
  maximum_n = 50,
  hv_genes = NULL,
  n.HVG = 3000,
  ...
)
}
\arguments{
\item{scExpr}{Single-cell expression data, with genes in rows and samples in columns}

\item{cell_type_labels}{a vector indicating cell-type level annotations}

\item{seurat_marker_method}{Seurat FindAllMarkers() test.use parameter. Default = 'wilcox'}

\item{log2FC}{log fold change threshold to select marker genes. Marker genes will be limited to a maximum of 'maximum_n' genes among those that pass the 'log2FC' threshold.}

\item{minimum_n}{minimum number of marker genes for a cell-type. If a cell type has fewer than 'minimum_n' genes passing the log2FC threshold, it will be excluded from the marker list}

\item{maximum_n}{maximum number of marker genes of a cell-type}

\item{hv_genes}{a character vector containing the names of high-variable genes. scExpr will be pre-filtered based on the provided hv_genes to reduce computation time during the differential expression (DE) analysis.
Set to NULL if not available and the function will automatically find hv genes using Seurat::FindVariableFeatures() function}

\item{n.HVG}{Number of features to select as top variable features: Seurat::FindVariableFeatures() function nfeatures parameter. Default to 3000}

\item{...}{additional parameters pass to Seurat::FindAllMarkers()}
}
\value{
a list of cell-type specific markers
}
\description{
Obtain cell-type specific markers using Seurat::FindAllMarkers()
}
