% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMarkers.R
\name{refMarkers_sigMatrixList}
\alias{refMarkers_sigMatrixList}
\title{Obtain cell-type specific markers from a list of signature matrices}
\usage{
refMarkers_sigMatrixList(sigMatrixList, maximum_n = 50)
}
\arguments{
\item{sigMatrixList}{a list of signature matrices}

\item{maximum_n}{maximum number of marker genes of a cell-type}
}
\value{
a list of markers obtained from different input signature matrices. Each list element is named after the corresponding signature matrix used to generate the markers.
}
\description{
Obtain cell-type specific markers from a list of signature matrices
}
