% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMatrix.R
\name{refMatrix}
\alias{refMatrix}
\title{Construct signature matrix from scRNA reference}
\usage{
refMatrix(
  methods,
  scExpr,
  cell_type_labels,
  cell_state_labels = NULL,
  hv_genes = NULL,
  log2FC = 2,
  minimum_n = 15,
  maximum_n = 50,
  order_by = "pval",
  max.spec_cutoff_for_DE = 0.3,
  markerList = NULL
)
}
\arguments{
\item{methods}{a character vector indicating which methods to use. Use list_refMarix() to check for available method names.}

\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}

\item{cell_state_labels}{a character vector indicating cell-states of each cells in scExpr. Pairwise DE analysis is initially conducted between cell states and then merges for cell types.
If set to NULL, it will default to the same values as the cell_type_labels. This is an argument required for 'scran' method only}

\item{hv_genes}{a character vector containing the names of high-variable genes. scExpr will be pre-filtered based on the provided hv_genes to reduce computation time during the differential expression (DE) analysis.
If set to NULL, the function will automatically select genes with specificity score passing 'max.spec_cutoff_for_DE' threshold as hv_genes. This argument is required for method 'limma' and 'scran'}

\item{log2FC}{log fold change threshold to select marker genes. Marker genes will be limited to a maximum of 'maximum_n' genes among those that pass the 'log2FC' threshold. This argument is required for method 'limma' and 'scran'}

\item{minimum_n}{minimum number of marker genes for a cell-type. This argument is required for method 'limma' and 'scran'}

\item{maximum_n}{maximum number of marker genes of a cell-type. This argument is required for method 'limma' and 'scran'}

\item{order_by}{a character specifying the criteria for ordering DE genes from scran analysis. Available options include 'pval' and 'min.lfc'. When more than 'maximum_n' genes pass the logFC threshold,
markers will be selected from the top 'maximum_n' genes ordered by this criterion. Default = 'pval'. This argument is required for 'scran' method only.}

\item{max.spec_cutoff_for_DE}{specificity score threshold to select for hv_genes. Default = 0.3. This argument is required for method 'limma' and 'scran'}

\item{markerList}{a list of pre-calculated cell-type marker list, where the first level of this list represents different methods used to derive cell-type specific markers, and the second level
comprises the actual cell-type specific genes identified by each method. This argument is required for 'markerList' method}
}
\value{
a list of signature matrices derived from scRNA reference
}
\description{
Generate a list of signature matrices from scRNA reference with user-selected methods
}
\examples{
\dontrun{
refMatrix(methods = c('raw','markerList'), scExpr = scExpr, cell_type_labels = scMeta$cell_type, markerList = markers)
}
}
