% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMatrix.R
\name{refMatrix_limma}
\alias{refMatrix_limma}
\title{Construct signature matrix using markers derived from limma DE analysis}
\usage{
refMatrix_limma(
  scExpr,
  cell_type_labels,
  hv_genes = NULL,
  log2FC = 2,
  log2FC_flexible = 1,
  minimum_n = 15,
  maximum_n = 50,
  max.spec_cutoff_for_DE = 0.3
)
}
\arguments{
\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}

\item{hv_genes}{a character vector containing the names of high-variable genes. scExpr will be pre-filtered based on the provided hv_genes to reduce computation time during the differential expression (DE) analysis.
If set to NULL, the function will automatically select genes with specificity score passing 'max.spec_cutoff_for_DE' threshold as hv_genes}

\item{log2FC}{log fold change threshold to select marker genes. Marker genes will be limited to a maximum of 'maximum_n' genes among those that pass the 'log2FC' threshold.}

\item{minimum_n}{minimum number of marker genes for a cell-type. If a cell type has fewer than 'minimum_n' genes passing the log2FC threshold, it will be excluded from the marker list}

\item{maximum_n}{maximum number of marker genes of a cell-type}

\item{max.spec_cutoff_for_DE}{specificity score threshold to select for hv_genes. Default = 0.3}
}
\value{
a signature matrix with genes in rows and cell-types in columns
}
\description{
The function first calculates the mean expression across cells from the same cell types, forming a reference matrix C. The C matrix is then filtered
based on markers genes identified from limma differential expression (DE) analysis.
}
