% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMatrix.R
\name{refMatrix_markerList}
\alias{refMatrix_markerList}
\title{Construct signature matrix with a pre-calculated marker list}
\usage{
refMatrix_markerList(scExpr, cell_type_labels, markerList)
}
\arguments{
\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}

\item{markerList}{a list of pre-calculated cell-type marker list, where the first level of this list represents different methods used
to derive cell-type specific markers, and the second level comprises the actual cell-type specific genes identified by each method}
}
\value{
a list of signature matrix
}
\description{
The function first calculates the mean expression across cells from the same cell types, forming a reference matrix C. The C matrix is then filtered
based on pre-calculated markers
}
