% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refMatrix.R
\name{refMatrix_raw}
\alias{refMatrix_raw}
\title{Construct signature matrix using average expression across cell types}
\usage{
refMatrix_raw(scExpr, cell_type_labels)
}
\arguments{
\item{scExpr}{Single-cell expression data to perform differential expression (DE) analysis, with genes in rows and samples in columns}

\item{cell_type_labels}{a character vector indicating cell-types of each cells in scExpr}
}
\value{
a signature matrix with genes in rows and cell-types in columns
}
\description{
This function calculates mean expression across cell types to form a signature matrix, without performing any gene filtering
}
