% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{simulator_averageCV}
\alias{simulator_averageCV}
\title{Average gene-set CV calculation for simulated and real bulk expression}
\usage{
simulator_averageCV(
  simulator_Statistics_res,
  genelist = NULL,
  msigdbr_species = "Homo sapiens",
  msigdbr_category = "H",
  msigdbr_subcategory = NULL
)
}
\arguments{
\item{simulator_Statistics_res}{a list of summary statistics retunred from simulator_Statistics() function}

\item{genelist}{a user-provided gene list. If set to NULL, will automatically retrieve gene sets collection from msigdbr}

\item{msigdbr_species}{msigdbr() species parameter. Required when genelist = NULL. Default = 'Homo sapiens'}

\item{msigdbr_category}{msigdbr() category parameter. Required when genelist = NULL. Default = 'H'}

\item{msigdbr_subcategory}{msigdbr() subcategory parameter.}
}
\value{
a dataframe of average CV for each gene set
}
\description{
Average gene-set CV calculation for simulated and real bulk expression
}
