% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{simulator_fracStatistics}
\alias{simulator_fracStatistics}
\title{Summarize all simulated fractions in simulated bulk object}
\usage{
simulator_fracStatistics(bulkSimulator_res, baseline_frac = NULL)
}
\arguments{
\item{bulkSimulator_res}{a list of simulated bulk object from bulkSimulator() function}

\item{baseline_frac}{fraction matrix in real bulk expression with cell types in columns and samples in rows. Set to NULL if not available}
}
\value{
a dataframe that documents all the simulated fractions within the simulated bulk object. This dataframe is particularly useful for downstream visualization purposes
}
\description{
Summarize all simulated fractions in simulated bulk object
}
