% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{simulator_Statistics}
\alias{simulator_Statistics}
\title{Calculate summary statistics for simulated and real bulk expression}
\usage{
simulator_Statistics(
  bulkSimulator_res,
  baseline_expr = NULL,
  var.genes = NULL,
  top_n_variable = 300,
  n.core = 1
)
}
\arguments{
\item{bulkSimulator_res}{a list of simulated bulk object from bulkSimulator() function}

\item{baseline_expr}{a real bulk expression matrix to compare against, with genes in rows and samples in columns}

\item{var.genes}{genes used to calculate gene-wise correlation and sample-wise correlation. If NULL, will automatically
choose top most variable genes from baseline_expr as var.genes}

\item{top_n_variable}{a numeric value indicating the number of top variable genes to be selected. Required when var.genes = NULL}

\item{n.core}{number of cores to use for parallel programming. Default = 1.}
}
\value{
a list of summary statistics for the simulated and real bulk expression. The list includes summary statistics such as
mean, variance, coefficient of variation (CV), Median Absolute Deviation (MAD), as well as gene-wise and sample-wise pairwise correlations.
}
\description{
Calculate summary statistics for simulated and real bulk expression
}
