% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_tcga_obj}
\alias{build_tcga_obj}
\title{Get tcga expression and purity estimates}
\usage{
build_tcga_obj(
  tcga_abbreviation,
  to_use_genes = NULL,
  purity_methods = c("ESTIMATE", "ABSOLUTE", "LUMP", "IHC", "CPE", "ABSOLUTE_GDC")
)
}
\arguments{
\item{tcga_abbreviation}{a character indicating tcga abbreviation for the tcga cohort to include, for example 'SKCM'}

\item{to_use_genes}{a vector indicating which genes to include for the exported tcga expression. Set to NULL to export full expression retrieved from xena browser}

\item{purity_methods}{a vector indicating tumor purity estimation method that is utilized as a means of estimating the malignant proportion within the exported object for TCGA expression data.
Available methods include 'ESTIMATE', 'ABSOLUTE', 'LUMP', 'IHC' and 'CPE' and 'ABSOLUTE_GDC' (ABSOLUTE_GDC contains ABSOLUTE downloaded from https://gdc.cancer.gov/about-data/publications/pancanatlas).
Default = c('ESTIMATE', 'ABSOLUTE', 'LUMP', 'IHC', 'CPE', 'ABSOLUTE_GDC'). Make sure you have the suggested package 'TCGAbiolinks' installed}
}
\value{
a list of tcga expression (named as simulated_bulk, TPM normalized and has hgnc symbols as rownames) and purity estimates (named as simulated_frac)
}
\description{
Given a tcga_abbreviation, download tcga expression from Xena browser and get purity estimates
}
\examples{
\dontrun{build_tcga_obj('SKCM')}

}
