% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_simulation.R
\name{bulkSimulator}
\alias{bulkSimulator}
\title{Bulk simulation function}
\usage{
bulkSimulator(
  methods,
  scExpr,
  scMeta,
  colnames_of_cellType = NA,
  colnames_of_subcluster = NA,
  colnames_of_sample = NA,
  simulated_frac = NULL,
  heter_cell_type = NA,
  ncells_perSample = 500,
  min_chunkSize = 5,
  use_chunk = "all",
  dirichlet_cs_par = 0.1,
  min.subcluster.size = 20,
  max.num.cs = NA,
  min.percentage = 1,
  max.percentage = 99,
  nbulk = 100,
  seed = 24,
  use_simulated_frac_as_prop_mat = F,
  disease = NULL,
  ct.sub = NULL,
  samplewithRep = T,
  n.core = 1
)
}
\arguments{
\item{methods}{a character vector indicating which bulk simulation methods to use. Use list_bulkSimulator() to check for available method names.}

\item{scExpr}{single cell expression matrix used to simulate bulk data, with genes in rows and cells in columns}

\item{scMeta}{dataframe that stores annotation info of each cells, rownames of scMeta should be equal to colnames of scExpr}

\item{colnames_of_cellType}{column name that corresponds to cellType in scMeta}

\item{colnames_of_subcluster}{column name that corresponds to subcluster info in scMeta, where subcluster contains sub-clustering information for each cellType.  This is an argument required for 'heter_sampleIDfree' method only.
Set to NA if subclustering information is not available; the function will then generate subclustering information automatically}

\item{colnames_of_sample}{column name that corresponds to sampleID in scMeta. Required for 'semi', 'heter', and 'SCDC' methods.}

\item{simulated_frac}{a matrix with pre-defined fraction of different cell types, with samples in rows and cell_types in columns. This argument if required for 'homo', 'semi', 'heter', 'heter_sampleIDfree', 'immunedeconv' methods}

\item{heter_cell_type}{name of the cell_type to maintain the highest level of heterogeneity. It is recommended to set this parameter to the name of the malignant cell types.
This argument is required for 'semi' and 'heter_sampleIDfree' methods}

\item{ncells_perSample}{number of cells to aggregate for each simulated bulk sample. This is an argument required for 'homo', 'semi', 'favilaco', 'immunedeconv' and 'SCDC' methods}

\item{min_chunkSize}{minimum number of cells to aggregate to construct a given cell-type component in the simulated bulk. This is an argument required for 'semi' and 'heter' methods}

\item{use_chunk}{a character indicating which cells to pool together for the a given cell_type. Default='all' other options include 'random'.
When use_chunk = 'all', use all the cells belonging to the same patient for a given cell type to generate the certain cell type component in the simulated bulk;
when use_chunk = 'random', randomly select 50-100\% of the cells belonging to the same patient for a given cell type. This is an argument required for 'semi' and 'heter' methods}

\item{dirichlet_cs_par}{a numeric value determine the dispersion level of the simulated fractions. With lower value indicating higher dispersion level. Default = 0.1. This is an argument required for 'heter_sampleIDfree' method.}

\item{min.subcluster.size}{minimum size of a subcluster. This argument is required when colnames_of_subcluster is NA. Default = 20. This is an argument for 'heter_sampleIDfree' method only.}

\item{max.num.cs}{maximum number of sub-clusters to aggregate for each cluster. If set to NA, this number will be equal to number of unique sub-clusters within each cell type. This is an argument for 'heter_sampleIDfree' method only.}

\item{min.percentage}{minimum percentage of cellType fraction to generate in fraction simulation. Default = 1. This argument is only required for 'favilaco'}

\item{max.percentage}{maximum percentage of cellType fraction to generate in fraction simulation. Default = 99. This argument is only required for 'favilaco'}

\item{nbulk}{number of simulated bulk samples. This argument is required for 'favilaco' and 'SCDC' methods.}

\item{seed}{a seed value for 'favilaco' method. Default = 24}

\item{use_simulated_frac_as_prop_mat}{a logical variable to determine whether to use 'simulated_frac' as 'prop_mat' as input for 'SCDC' method}

\item{disease}{indicate the health condition of subjects. This argument is only required for 'SCDC' method.}

\item{ct.sub}{a subset of cell types that are selected to construct pseudo bulk samples. If NULL, then all cell types are used. This argument is only required for 'SCDC' method.}

\item{samplewithRep}{logical, randomly sample single cells with replacement. Default is T. This argument is only required for 'SCDC' method.}

\item{n.core}{number of cores to use for parallel programming. Default = 1. Parallel programming can be enabled for 'homo', 'semi' and 'heter' methods.}
}
\value{
a list of simulated bulk objects
}
\description{
Generate a list of bulk simulation objects using user-selected simulation methods
}
\examples{
\dontrun{
# first create a fraction matrix for the simulated bulk
simulated_frac = fracSimulator_Beta(scMeta, n = 100,
                                    colnames_of_sample = 'sampleID',
                                    colnames_of_cellType = 'cell_type',
                                    fixed_cell_type = 'malignant')

bulkSimulator(methods = c('homo', 'semi','heter','favilaco','immunedeconv','SCDC'),
              scExpr = scExpr,
              scMeta = scMeta,
              colnames_of_cellType = 'cell_type',colnames_of_sample = 'sampleID',

              # pre-defined simulated_frac will be used for 'homo', 'semi, 'heter' and 'immunedeconv' methods
              simulated_frac = simulated_frac,

              # argument for semi/heter_sampleIDfree bulk simulation method
              heter_cell_type = 'malignant',

              # argument for heter_sampleIDfree bulk simulation method
              dirichlet_cs_par = 0.1,
              min.subcluster.size = 20,

              # argument for 'favilaco' and 'SCDC' method
              nbulk = 100,
              n.core = 16)
}
}
