% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{plot_PairwiseGeneCorrelation}
\alias{plot_PairwiseGeneCorrelation}
\title{Plot pairwise gene correlation in simulated bulk and real bulk expression}
\usage{
plot_PairwiseGeneCorrelation(
  simulator_Statistics_res,
  nrow_panel = 2,
  desired_order = c("baseline", "homo", "semi", "heter", "heter_sampleIDfree",
    "favilaco", "immunedeconv", "SCDC"),
  ...
)
}
\arguments{
\item{simulator_Statistics_res}{a list of summary statistics returned from simulator_Statistics() function}

\item{nrow_panel}{nrow of the ggplot panels}

\item{desired_order}{a character vector indicating desired order to organize simulated bulk in the figure.
The default order is specified as c('baseline','homo','semi','heter','heter_sampleIDfree','favilaco','immunedeconv','SCDC')}

\item{...}{additional arguments to pass to geom_tile()}
}
\value{
a panel of heatmaps showing pairwise gene correlations
}
\description{
Plot pairwise gene correlation in simulated bulk and real bulk expression
}
