% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodInspector.R
\name{plot_PairwiseSampleCorrelation}
\alias{plot_PairwiseSampleCorrelation}
\title{Plot sample wise correlations in simulated bulk and real bulk}
\usage{
plot_PairwiseSampleCorrelation(simulator_Statistics_res, ...)
}
\arguments{
\item{simulator_Statistics_res}{a list of summary statistics returned from simulator_Statistics() function}

\item{...}{additional arguments to pass to theme()}
}
\value{
a panel of boxplot showing the distribution of sample-wise correlations
}
\description{
Plot sample wise correlations in simulated bulk and real bulk
}
